Prepare 0.11.0 release
Merging in a huge release from dropshot origin.

The highlight of this release is that they have created a trait based
way to define the API structure and generate the OpenAPI code. This
makes it harder to get into cycles where you cannot fix the API code
because the generated code is broken and you cannot fix the generated
code because the API wont build. Which was a problem with the function
based alternative.

Other added features can be found on their [changelog here.](https://github.com/oxidecomputer/dropshot/blob/main/CHANGELOG.adoc#0-11-0-released-2024-08-21)